+++ CHANGELOG.md
@@ -97,6 +97,8 @@ We change btn colors:
 - color="success"
 + color="air-primary-success"
 
+* fix(Button/Link): merge `active-class` / `inactive-class` with app config (#4446)
+
 ## [0.7.2](https://github.com/bitrix24/b24ui/compare/v0.7.1...v0.7.2) (2025-07-14)
 
 ### Bug Fixes